+++ jstests/tests/user/mock_jsapi.js
