+++ jstests/testsuite.js
@@ -1,7 +1,7 @@
-require("./tests/home/test_configurables.js");
-require("./tests/home/test_models.js");
-require("./tests/home/test_application_list_view.js");
-require("./tests/home/test_application_view.js");
+require("./tests/user/test_configurables.js");
+require("./tests/user/test_models.js");
+require("./tests/user/test_application_list_view.js");
+require("./tests/user/test_application_view.js");
 require("./tests/vue/components/test_DataTable.js");
 require("./tests/vue/components/test_ConfirmDialog.js");
 require("./tests/test_utils.js");
