+++ remoteappmanager/application.py
@@ -2,7 +2,7 @@ from tornado import web
 
 from remoteappmanager.base_application import BaseApplication
 from remoteappmanager.handlers.api import (
-    HomeHandler, RegisterContainerHandler)
+    UserHomeHandler, RegisterContainerHandler)
 from remoteappmanager.utils import url_path_join, without_end_slash
 from remoteappmanager import webapi
 
@@ -20,7 +20,7 @@ class Application(BaseApplication):
             (without_end_slash(
                 url_path_join(base_urlpath, "containers", "([a-z0-9_]*)")
             )+"/?", RegisterContainerHandler),
-            (base_urlpath, HomeHandler),
+            (base_urlpath, UserHomeHandler),
             (base_urlpath.rstrip('/'), web.RedirectHandler, {
                 "url": base_urlpath}),
         ]
