+++ remoteappmanager/handlers/api.py
@@ -1,4 +1,4 @@
-from .home_handler import HomeHandler  # noqa
+from .user_home_handler import UserHomeHandler  # noqa
 from .base_handler import BaseHandler  # noqa
 from .register_container_handler import RegisterContainerHandler  # noqa
 from .admin.admin_home_handler import AdminHomeHandler  # noqa
