+++ remoteappmanager/handlers/tests/test_user_home_handler.py
@@ -5,7 +5,7 @@ from remoteappmanager.tests.mocking import dummy
 from remoteappmanager.tests.temp_mixin import TempMixin
 
 
-class TestHomeHandler(TempMixin, utils.AsyncHTTPTestCase, LogTrapTestCase):
+class TestUserHomeHandler(TempMixin, utils.AsyncHTTPTestCase, LogTrapTestCase):
     def get_app(self):
         app = dummy.create_application()
         app.hub.verify_token.return_value = {
