+++ remoteappmanager/handlers/user_home_handler.py
@@ -3,10 +3,10 @@ from tornado import gen, web
 from remoteappmanager.handlers.base_handler import BaseHandler
 
 
-class HomeHandler(BaseHandler):
+class UserHomeHandler(BaseHandler):
     """Render the user's home page"""
 
     @web.authenticated
     @gen.coroutine
     def get(self):
-        self.render('home.html')
+        self.render('user/page.html')
