+++ remoteappmanager/static/js/user/configurables.js
