+++ remoteappmanager/static/js/user/main.js
