+++ remoteappmanager/static/js/user/user-resources.js
