+++ remoteappmanager/static/js/user/views/application_list_view.js
