+++ remoteappmanager/static/js/user/views/application_view.js
