+++ remoteappmanager/templates/user/page.html
